Change cli arguments
